social_media_automation/requirements.txt
google-generativeai==0.8.3
google-cloud-texttospeech==2.16.3
google-cloud-speech==2.26.1
moviepy==1.0.3
requests==2.32.3
python-dotenv==1.0.1
